+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Attach/detach policy to user/group (see [wiki examples](https://github.com/wallix/awless/wiki/Examples))
 - Attach/detach user to group (see [wiki examples](https://github.com/wallix/awless/wiki/Examples))
+- List AWS load balancers with `awless list loadbalancers`
 
 ### Bugfixes